+++ cluster_agent/settings.py
@@ -16,7 +16,6 @@ class Settings(BaseSettings):
     # slurmrestd info
     BASE_SLURMRESTD_URL: str = Field("http://127.0.0.1:6820", regex=_URL_REGEX)
     X_SLURM_USER_NAME: str = Field("root")
-    X_SLURM_USER_TOKEN: Optional[str]
     DEFAULT_SLURM_WORK_DIR: str = Field("/tmp")
 
     # cluster api info