+++ tests/identity/test_slurmrestd_identity.py
@@ -118,9 +118,7 @@ def test__load_token_from_cache__returns_none_cached_token_will_expire_soon(
     assert retrieved_token is None
 
 
-def test_acquire_token__gets_a_token_from_the_cache(
-    mock_slurmrestd_api_cache_dir, tweak_settings
-):
+def test_acquire_token__gets_a_token_from_the_cache(mock_slurmrestd_api_cache_dir):
     """
     Verifies that the token is retrieved from the cache if it is found there.
     """
@@ -133,8 +131,7 @@ def test_acquire_token__gets_a_token_from_the_cache(
         algorithm="HS256",
     )
     token_path.write_text(created_token)
-    with tweak_settings(X_SLURM_USER_TOKEN=None):
-        retrieved_token = acquire_token("dummy-user")
+    retrieved_token = acquire_token("dummy-user")
     assert retrieved_token == created_token
 
 
@@ -142,7 +139,6 @@ def test_acquire_token__gets_a_token_from_the_cache(
 def test_acquire_token__gets_a_token_from_slurm_if_one_is_not_in_the_cache(
     mock_subprocess_popen,
     mock_slurmrestd_api_cache_dir,
-    tweak_settings,
 ):  # noqa
     """
     Verifies that a token is pulled from Slurm if it is not found in the cache.
@@ -161,8 +157,7 @@ def test_acquire_token__gets_a_token_from_slurm_if_one_is_not_in_the_cache(
 
     mock_subprocess_popen.return_value = process_mock
 
-    with tweak_settings(X_SLURM_USER_TOKEN=None):
-        retrieved_token = acquire_token("dummy-user")
+    retrieved_token = acquire_token("dummy-user")
     assert retrieved_token == "dummy-token"
 
     token_path = mock_slurmrestd_api_cache_dir / "dummy-user.token"
@@ -173,7 +168,6 @@ def test_acquire_token__gets_a_token_from_slurm_if_one_is_not_in_the_cache(
 def test_acquire_token__raise_error_if_subprocess_command_failed(
     mock_subprocess_popen,
     mock_slurmrestd_api_cache_dir,
-    tweak_settings,
 ):  # noqa
     """
     Verifies whether an error is raised or not in case "scontrol token" subprocess call fails.
@@ -191,8 +185,7 @@ def test_acquire_token__raise_error_if_subprocess_command_failed(
 
     mock_subprocess_popen.return_value = process_mock
 
-    with tweak_settings(X_SLURM_USER_TOKEN=None):
-        with pytest.raises(ProcessExecutionError) as error:
-            acquire_token("dummy-user")
+    with pytest.raises(ProcessExecutionError) as error:
+        acquire_token("dummy-user")
 
     assert "This is a dummy error message" in str(error.value)